Add show to graffiti controller & refactor angular
The graffiti now has a show page where the json for the graffitis
information can be found/viewed/requested.

The angular has also been refactored to http request the graffiti json
from the newly implemented show page, this will help in the future with
the edit page and serve as the basis if the angular ever needs more
than just the grid information.