Refactor where angular gets the starting graffiti
The angular originally obtained the starting information about the
graffiti by making a get request, this was the simplest way at the
time.

But since then the angular now has an ng-init which allowed it to
run a function using the id, the graffiti.drawing is now also fed into
this function and is assigned at that point. Tests and views have been
altered appropriately.