Apply the wall to the homepage
The wall now displays a button for each graffiti that will be shown on
a wall. These buttons will be used if a user want to edit/ overwrite
existing graffiti.

Initially the angular was going to be fed the graffiti drawings
directly but that was too much information at once. To try and reduce
the amount of info the graffiti drawings have been compressed, but
this was still too much time so this idea was scrapped. The graffiti
will be fetched using get requests at a later point, but the
compression was kept.