Apply method for assigning graffiti to the wall
A method was created to search through the wall to find appropriate
places graffiti can go, it starts at the center and expands outwards in
a diamond formation where it randomly selects a valid location for the
graffiti.

This method has been applied in the controllers so the graffiti is now
no longer just placed in slot 11,11.